refactor(dashboard/admin-user-balance): moved component to users folder in admin users module
