+++ apps/dashboard/src/locales/en/modules/admin.json
@@ -1,0 +1,44 @@
+{
+  "userOverview": {
+    "title": "User overview",
+    "list": {
+      "header": "User overview"
+    }
+  },
+  "singleUser": {
+    "title": "User overview",
+    "profile": "{user}'s profile",
+    "userInfo": {
+      "header": "User information"
+    },
+    "balance": {
+      "header": "Balance"
+    }
+  },
+  "banners": {
+    "title": "Banners",
+    "list": {
+      "header": "Banner overview",
+      "filterActive": "Filter active",
+      "filterNotExpired": "Filter not expired",
+      "active": "Active",
+      "notActive": "Not active"
+    }
+  },
+  "forms": {
+    "user":{
+      "header": "Create user",
+      "ofAge": "User is of age",
+      "active": "User is active"
+    },
+    "banner": {
+      "headerCreate": "Create banner",
+      "headerEdit": "Edit banner",
+      "noBannerFound": "No banner found, please upload one!",
+      "duration": "Duration",
+      "timeSpan": "Timespan",
+      "till": "till",
+      "active": "Active"
+    }
+  }
+}
